+++ src/applications/vaos/appointment-list/pages/UpcomingAppointmentsDetailsPage/index.jsx
@@ -3,7 +3,7 @@ import { shallowEqual, useDispatch, useSelector } from 'react-redux';
 import { useParams } from 'react-router-dom';
 import moment from 'moment';
 import InfoAlert from '../../../components/InfoAlert';
-import AppointmentDetailsErrorMessage from '../../../components/AppointmentDetailsErrorMessage';
+import AppointmentDetailsErrorMessage from '../../components/AppointmentDetailsErrorMessage';
 import FullWidthLayout from '../../../components/FullWidthLayout';
 import CCLayout from '../../../components/layouts/CCLayout';
 import VideoLayout from '../../../components/layouts/VideoLayout';
